Module zcash_extensions::consensus::transparent

Consensus logic for Transparent Zcash Extensions.

Structs

Context

The complete set of context data that is available to any extension having an assigned extension type ID. This type may be modified in the future if additional context information is required by newly integrated TZEs.

InvalidExtId

Enums

ExtensionId

The set of programs that have assigned type IDs within the Zcash consensus rules.

Constants

EXTENSION_DEMO

Wire value for the demo extension identifier.

Traits

Epoch

Implementations of this trait provide complete extension validation rules for a specific epoch, and handle dispatch of verification to individual TZEs based upon extension ID and mode.

Functions

epoch_for_branch