Trait zcash_extensions::consensus::transparent::Epoch

pub trait Epoch {
    type Error;
    fn verify<'a>(
        &self, 
        precondition: &Precondition, 
        witness: &Witness<AuthData>, 
        ctx: &Context<'a>
    ) -> Result<(), Error<Self::Error>>;
}

Implementations of this trait provide complete extension validation rules for a specific epoch, and handle dispatch of verification to individual TZEs based upon extension ID and mode.

Associated Types

type Error

Required methods

fn verify<'a>(
&self,
precondition: &Precondition,
witness: &Witness<AuthData>,
ctx: &Context<'a>
) -> Result<(), Error<Self::Error>>

For a specific epoch, if the extension ID and mode of the supplied witness matches that of the supplied precondition, these values will be passed to the associated extension for verification, along with whatever that extension requires of the provided Context.

Successful validation is indicated by the returned Result containing no errors.