Module zcash_extensions::transparent::demo

Demo implementation of TZE consensus rules.

The demo program implements a dual-hash-lock encumbrance with the following form:

hash = BLAKE2b_256(preimage_1 || BLAKE2b_256(preimage_2))

The two preimages are revealed in sequential transactions, demonstrating how TZEs can impose constraints on how program modes are chained together.

The demo program has two modes:

Mode 0: hash_1 = BLAKE2b_256(preimage_1 || hash_2)

Mode 1: hash_2 = BLAKE2b_256(preimage_2)

and uses the following transaction formats:

tx_a: [ [any input types...] ----> TzeOut(value, hash_1) ]

tx_b: [ TzeIn(tx_a, preimage_1) -> TzeOut(value, hash_2) ]

tx_c: [ TzeIn(tx_b, preimage_2) -> [any output types...] ]

Structs

DemoBuilder

Wrapper for zcash_primitives::transaction::builder::Builder that simplifies constructing transactions that utilize the features of the demo extension.

Program

Marker type for the demo extension.

Enums

DemoBuildError

Errors that can occur in construction of transactions using DemoBuilder.

Error

Errors that may be produced during parsing and verification of demo preconditions and witnesses.

Precondition

The precondition type for the demo extension.

Witness

The witness type for the demo extension.

Traits

Context

This trait defines the context information that the demo extension requires be made available to it by a consensus node integrating this extension.