Trait zcash_extensions::transparent::demo::Context

pub trait Context {
    fn is_tze_only(&self) -> bool;

    fn tx_tze_outputs(&self) -> &[TzeOut];
}

This trait defines the context information that the demo extension requires be made available to it by a consensus node integrating this extension.

This context type provides accessors to information relevant to a single transaction being validated by the extension.

Required methods

fn is_tze_only(&self) -> bool

Predicate used to determine whether this transaction has only TZE inputs and outputs. The demo extension does not support verification of transactions which have either shielded or transparent inputs and outputs.

fn tx_tze_outputs(&self) -> &[TzeOut]

List of all TZE outputs in the transaction being validate by the extension.